backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
drf-yasg==1.21.7
mysqlclient==2.2.1
redis==5.0.1
celery==5.3.6
gunicorn==21.2.0
python-dotenv==1.0.1
requests==2.31.0
PyJWT==2.8.0
pytest==8.0.2
pytest-django==4.8.0
coverage==7.4.1
Faker==22.6.0
black==24.2.0
flake8==7.0.0
mypy==1.8.0
requests-mock